Update patch set 3
Patch Set 3:

From the time being, the easiest way is to put it in the docker dir and then call COPY in Dockerfile to replace Functest's logging.ini(like testcases.yaml).

Then it's useless to init logging in the testcase as it's done via run_tests.py

logging.ini seems fine.

Patch-set: 3