Update patch set 5
Patch Set 5:

> I consider that we shouldn't describe any testcase in Functest if
 > it's not tested by OPNFV gating. Else any enduser is free to update
 > the testcases.yaml locally.
The project of clover_k8s configured in testcases.yaml is Clover instead of Functest, so in CI it would be tested by OPNFV gating, since no matter it fails or not, the result of k8s job only depends on k8s_smoke and k8s_conformance.

 > Here we should stop listing all K8s installers. We could simply
 > exclude os- scenarios instead.
Only using "k8" as scenario to constrain is not enough, because clover and stor4nfv should not run in case k8-nosdn-nofeature scenario, except testcases.yaml is updated by users locally.

Patch-set: 5