Deploy Functest Kubernetes CI toolchains
It eases verifying OpenStack by deploying the full testing toolchain
via Ansible. As opposed to the Releng model, all tests can be executed
in parallel.

It leverages on Xtesting Ansible role which currently deploys
Jenkins, Minio, S3www, MongoDB, TestAPI and loads all jobs in Jenkins.
The dashboard will be deployed in the next update of the ansible role.

To deploy the testing toolchain:
$ ansible-galaxy install collivier.xtesting
$ ansible-playbook ansible/site.yml

Change-Id: Ie409653ae4f7f4d315449ee54ae0c504272f7899
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>