Update patch set 1
Patch Set 1:

I think in case of daemonset, by default, system attempts to launch a pod in every node. In this case there are three nodes in cluster and in one of nodes the pod didn't get up and running. In that sense I would say the test is working correctly but one of the nodes is not.

Or is 'control-plane' a node where pods are not expected to be scheduled at all? In Kubernetes.create_check_and_delete_daemonset one can define which nodes to filter (node_labels).

Patch-set: 1